Split the blacklist tests, move Blacklist to better package
