+++ SpiNNaker-comms/src/test/resources/uk/ac/manchester/spinnaker/alloc/bmp/example.blacklist
